requirements.txt
pytelegrambotapi
websocket-client
markdownify